Update script.js
not sure lines 55 and 56 are needed (given xOffset and yOffset are already re-assigned during drag function).  The redundancy is a little confusing when trying to follow the solution, so wanted to suggest perhaps considering removing those reassignments of initialX and initialY on lines 55-56 since they get overwritten anyhow in dragStart function (lines 18 and 19).